--- components/sheet/gorhom-sheet.tsx
@@ -1,3 +1,0 @@
-import BottomSheet from '@gorhom/bottom-sheet'
-
-export { BottomSheet as GorhomSheet }
