+++ eslint.config.js
@@ -1,10 +1,15 @@
 // https://docs.expo.dev/guides/using-eslint/
-const { defineConfig } = require('eslint/config');
-const expoConfig = require('eslint-config-expo/flat');
+const { defineConfig } = require('eslint/config')
+const expoConfig = require('eslint-config-expo/flat')
 
 module.exports = defineConfig([
   expoConfig,
   {
     ignores: ['dist/*'],
   },
-]);
+  {
+    rules: {
+      'import/namespace': 'off',
+    },
+  },
+])
