+++ src/app/_layout.tsx
@@ -1,0 +1,34 @@
+import { Providers } from '@/providers/Providers'
+import '@/styles/globals.css'
+import { useFonts } from 'expo-font'
+import { SplashScreen, Stack } from 'expo-router'
+import { useEffect } from 'react'
+
+export default function RootLayout() {
+  const [fontsLoaded, error] = useFonts({
+    'Pretendard-Thin': require('@/assets/fonts/Pretendard-Thin.otf'),
+    'Pretendard-ExtraLight': require('@/assets/fonts/Pretendard-ExtraLight.otf'),
+    'Pretendard-Light': require('@/assets/fonts/Pretendard-Light.otf'),
+    'Pretendard-Regular': require('@/assets/fonts/Pretendard-Regular.otf'),
+    'Pretendard-Medium': require('@/assets/fonts/Pretendard-Medium.otf'),
+    'Pretendard-SemiBold': require('@/assets/fonts/Pretendard-SemiBold.otf'),
+    'Pretendard-Bold': require('@/assets/fonts/Pretendard-Bold.otf'),
+    'Pretendard-ExtraBold': require('@/assets/fonts/Pretendard-ExtraBold.otf'),
+    'Pretendard-Black': require('@/assets/fonts/Pretendard-Black.otf'),
+  })
+
+  useEffect(() => {
+    if (error) throw error
+    if (fontsLoaded) SplashScreen.hideAsync()
+  }, [fontsLoaded, error])
+
+  if (!fontsLoaded) return null
+
+  return (
+    <Providers>
+      <Stack>
+        <Stack.Screen name="index" options={{ headerShown: false }} />
+      </Stack>
+    </Providers>
+  )
+}
