+++ src/styles/zIndex.ts
@@ -1,3 +1,4 @@
 export const zIndex = {
   modal: 1000,
+  dim: 999,
 } as const
